services/books/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
pydantic
redis
httpx
python-jose
passlib[bcrypt]
python-multipart
uuid